requirements.txt
python-dotenv~=1.0.1
cachetools~=5.4.0
pytz~=2024.1

# setuptools~=70.2.0
